requirements.txt
amqp==5.2.0
APScheduler==3.10.4
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.3
Django==5.0.7
django-allauth==0.63.6
django-apscheduler==0.6.2
django-background-tasks==1.2.5
django-celery-beat==2.6.0
django-compat==1.0.15
django-crontab==0.7.1
django-timezone-field==7.0
et-xmlfile==1.1.0
humanize==4.10.0
kombu==5.3.7
numpy==2.0.0
openpyxl==3.1.5
pandas==2.2.2
pillow==10.4.0
prompt_toolkit==3.0.47
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.7
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
tzlocal==5.2
vine==5.1.0
wcwidth==0.2.13
